# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "MusicRequest": {
        "additionalProperties": false,
        "example": {
          "input": [
            {
              "text": "an upbeat pop song about summer",
              "type": "text"
            }
          ],
          "model": "gateway-music-lyria",
          "parameters": {
            "bpm": 120,
            "duration_seconds": 30,
            "file_format": "wav"
          }
        },
        "properties": {
          "input": {
            "description": "Non-empty ordered music-generation inputs.",
            "items": {
              "discriminator": {
                "mapping": {
                  "audio": "#/components/schemas/MusicAudioInput",
                  "image": "#/components/schemas/MusicImageInput",
                  "lyrics": "#/components/schemas/LyricsInput",
                  "text": "#/components/schemas/TextInput"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextInput"
                },
                {
                  "$ref": "#/components/schemas/LyricsInput"
                },
                {
                  "$ref": "#/components/schemas/MusicImageInput"
                },
                {
                  "$ref": "#/components/schemas/MusicAudioInput"
                }
              ]
            },
            "minItems": 1,
            "title": "Input",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "description": "Client-owned metadata returned unchanged with the task.",
            "title": "Metadata",
            "type": "object"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model id returned by GET /v1/models, or omit / set to `auto` to let the gateway auto-route to a backend whose limits fit the request's input (modalities, dimensions, duration, ...).",
            "title": "Model"
          },
          "parameters": {
            "$ref": "#/components/schemas/MusicParameters"
          },
          "routing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutingDirective"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "input"
        ],
        "title": "MusicRequest",
        "type": "object"
      },
      "MusicTaskResponse": {
        "additionalProperties": true,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "links": {
            "$ref": "#/components/schemas/ResourceLinks"
          },
          "lyrics": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lyrics"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "object": {
            "const": "music",
            "default": "music",
            "title": "Object",
            "type": "string"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/MusicOutput"
            },
            "title": "Outputs",
            "type": "array"
          },
          "routing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed",
              "cancelled",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "model",
          "status",
          "created_at",
          "links"
        ],
        "title": "MusicTaskResponse",
        "type": "object"
      },
      "ProblemDetail": {
        "additionalProperties": true,
        "description": "RFC 9457 problem details with stable gateway extensions.",
        "properties": {
          "code": {
            "description": "Stable machine-readable gateway error code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable detail for this occurrence.",
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Field-level validation errors, when applicable.",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI reference identifying this occurrence.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request correlation identifier.",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP response status code.",
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "API key",
        "description": "Front-end API key sent as \"Authorization: Bearer <token>\".",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Unified image, video, and music gateway with separate REST APIs.",
    "title": "mm-gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/music": {
      "post": {
        "operationId": "createMusic",
        "parameters": [
          {
            "description": "Client-generated key used to safely retry this create request. Reuse with a different body returns 409.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client-generated key used to safely retry this create request. Reuse with a different body returns 409.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MusicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2026-08-09T12:00:00Z",
                  "id": "mus_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                  "links": {
                    "self": "https://gateway.example.test/v1/music/mus_01HZX4J3K7NQ8X2V9Y6R5W4T3P"
                  },
                  "metadata": {},
                  "model": "gateway-music-lyria",
                  "object": "music",
                  "outputs": [],
                  "status": "pending"
                },
                "schema": {
                  "$ref": "#/components/schemas/MusicTaskResponse"
                }
              }
            },
            "description": "The music task was accepted.",
            "headers": {
              "ETag": {
                "description": "Version identifier for conditional polling.",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "description": "true when the response replays an earlier create request.",
                "schema": {
                  "type": "boolean"
                }
              },
              "Link": {
                "description": "Canonical task URL with rel=\"self\".",
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "description": "Canonical URL of the created task resource.",
                "schema": {
                  "format": "uri",
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Suggested number of seconds before polling again.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Invalid request",
                    "value": {
                      "code": "invalid_request_error",
                      "detail": "The routing profile is unavailable.",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 400,
                      "title": "Invalid Request Error",
                      "type": "urn:mm-gateway:problem:invalid_request_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Invalid request (invalid_request_error / unsupported_feature)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Missing API key",
                    "value": {
                      "code": "unauthorized",
                      "detail": "Missing API key. Send an 'Authorization: Bearer <token>' header.",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 401,
                      "title": "Unauthorized",
                      "type": "urn:mm-gateway:problem:unauthorized"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Missing or unknown API key (unauthorized)."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Forbidden",
                    "value": {
                      "code": "forbidden",
                      "detail": "The API key is not allowed to perform this request.",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 403,
                      "title": "Forbidden",
                      "type": "urn:mm-gateway:problem:forbidden"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Key not allowed to perform the request (forbidden)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Unknown task",
                    "value": {
                      "code": "task_not_found",
                      "detail": "image task img_unknown not found",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 404,
                      "title": "Task Not Found",
                      "type": "urn:mm-gateway:problem:task_not_found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Model or task not found."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Idempotency conflict",
                    "value": {
                      "code": "idempotency_conflict",
                      "detail": "The Idempotency-Key was already used with a different request.",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 409,
                      "title": "Idempotency Conflict",
                      "type": "urn:mm-gateway:problem:idempotency_conflict"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Idempotency key conflicts with an earlier request."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Validation failed",
                    "value": {
                      "code": "validation_error",
                      "detail": "Request validation failed.",
                      "errors": [
                        {
                          "loc": [
                            "body",
                            "model"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ],
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 422,
                      "title": "Validation Error",
                      "type": "urn:mm-gateway:problem:validation_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service error",
                    "value": {
                      "code": "generation_service_error",
                      "detail": "The generation service returned an error.",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 502,
                      "title": "Generation Service Error",
                      "type": "urn:mm-gateway:problem:generation_service_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service returned an error."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service unavailable",
                    "value": {
                      "code": "generation_service_unavailable",
                      "detail": "No generation service is available for this request.",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 503,
                      "title": "Generation Service Unavailable",
                      "type": "urn:mm-gateway:problem:generation_service_unavailable"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No usable generation service is configured."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service timeout",
                    "value": {
                      "code": "generation_service_timeout",
                      "detail": "The generation service timed out.",
                      "instance": "/v1/music",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 504,
                      "title": "Generation Service Timeout",
                      "type": "urn:mm-gateway:problem:generation_service_timeout"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service timed out."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a music task",
        "tags": [
          "music"
        ]
      }
    },
...
    "/v1/music/{music_id}": {
      "get": {
        "operationId": "getMusic",
        "parameters": [
          {
            "description": "Opaque music task id.",
            "in": "path",
            "name": "music_id",
            "required": true,
            "schema": {
              "description": "Opaque music task id.",
              "title": "Music Id",
              "type": "string"
            }
          },
          {
            "description": "Previously returned ETag; unchanged resources return 304.",
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Previously returned ETag; unchanged resources return 304.",
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "completed_at": "2026-08-09T12:00:30Z",
                  "created_at": "2026-08-09T12:00:00Z",
                  "id": "mus_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                  "links": {
                    "self": "https://gateway.example.test/v1/music/mus_01HZX4J3K7NQ8X2V9Y6R5W4T3P"
                  },
                  "lyrics": "[verse]\nWalking down the street...",
                  "metadata": {},
                  "model": "gateway-music-lyria",
                  "object": "music",
                  "outputs": [
                    {
                      "mime_type": "audio/wav",
                      "uri": "https://cdn.example.test/mus/abc.wav"
                    }
                  ],
                  "status": "succeeded",
                  "usage": {
                    "cost": 0.05,
                    "duration_seconds": 30,
                    "output_count": 1
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/MusicTaskResponse"
                }
              }
            },
            "description": "The latest music task state.",
            "headers": {
              "ETag": {
                "description": "Version identifier for conditional polling.",
                "schema": {
                  "type": "string"
                }
              },
              "Link": {
                "description": "Canonical task URL with rel=\"self\".",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Suggested number of seconds before polling again.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "304": {
            "description": "The task representation has not changed."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Invalid request",
                    "value": {
                      "code": "invalid_request_error",
                      "detail": "The routing profile is unavailable.",
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 400,
                      "title": "Invalid Request Error",
                      "type": "urn:mm-gateway:problem:invalid_request_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Invalid request (invalid_request_error / unsupported_feature)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Missing API key",
                    "value": {
                      "code": "unauthorized",
                      "detail": "Missing API key. Send an 'Authorization: Bearer <token>' header.",
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 401,
                      "title": "Unauthorized",
                      "type": "urn:mm-gateway:problem:unauthorized"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Missing or unknown API key (unauthorized)."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Forbidden",
                    "value": {
                      "code": "forbidden",
                      "detail": "The API key is not allowed to perform this request.",
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 403,
                      "title": "Forbidden",
                      "type": "urn:mm-gateway:problem:forbidden"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Key not allowed to perform the request (forbidden)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Unknown task",
                    "value": {
                      "code": "task_not_found",
                      "detail": "image task img_unknown not found",
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 404,
                      "title": "Task Not Found",
                      "type": "urn:mm-gateway:problem:task_not_found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unknown music task id."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Validation failed",
                    "value": {
                      "code": "validation_error",
                      "detail": "Request validation failed.",
                      "errors": [
                        {
                          "loc": [
                            "body",
                            "model"
                          ],
                          "msg": "Field required",
                          "type": "missing"
                        }
                      ],
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 422,
                      "title": "Validation Error",
                      "type": "urn:mm-gateway:problem:validation_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service error",
                    "value": {
                      "code": "generation_service_error",
                      "detail": "The generation service returned an error.",
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 502,
                      "title": "Generation Service Error",
                      "type": "urn:mm-gateway:problem:generation_service_error"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service returned an error."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service unavailable",
                    "value": {
                      "code": "generation_service_unavailable",
                      "detail": "No generation service is available for this request.",
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 503,
                      "title": "Generation Service Unavailable",
                      "type": "urn:mm-gateway:problem:generation_service_unavailable"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No usable generation service is configured."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "error": {
                    "summary": "Generation service timeout",
                    "value": {
                      "code": "generation_service_timeout",
                      "detail": "The generation service timed out.",
                      "instance": "/v1/music/{music_id}",
                      "request_id": "req_01HZX4J3K7NQ8X2V9Y6R5W4T3P",
                      "status": 504,
                      "title": "Generation Service Timeout",
                      "type": "urn:mm-gateway:problem:generation_service_timeout"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Generation service timed out."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retrieve a music task",
        "tags": [
          "music"
        ]
      }
    },
...
  }
}